added phpstan rule for monorepo to avoid typehints in entity setter parameters
